+++ ts/domlib/surround/append-merge.ts
@@ -1,0 +1,43 @@
+// Copyright: Ankitects Pty Ltd and contributors
+// License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
+
+import type { ParseFormat } from "./parse-format";
+import type { TreeNode } from "./tree-node";
+import { FormattingNode } from "./tree-node";
+
+/**
+ * @param last: The node which is intended to be the new end node.
+ */
+function mergeInNode(
+    initial: TreeNode[],
+    last: FormattingNode,
+    format: ParseFormat,
+): TreeNode[] {
+    const minimized: TreeNode[] = [last];
+
+    for (let i = initial.length - 1; i >= 0; i--) {
+        const next = initial[i];
+
+        let merged: FormattingNode | null;
+        if (next instanceof FormattingNode && (merged = format.tryMerge(next, last))) {
+            minimized[0] = merged;
+        } else {
+            minimized.unshift(...initial.slice(0, i + 1));
+            break;
+        }
+    }
+
+    return minimized;
+}
+
+export function appendNode(
+    nodes: TreeNode[],
+    node: TreeNode,
+    format: ParseFormat,
+): TreeNode[] {
+    if (node instanceof FormattingNode) {
+        return mergeInNode(nodes, node, format);
+    } else {
+        return [...nodes, node];
+    }
+}
