--- ts/domlib/surround/find-adjacent.test.ts
@@ -1,149 +1,0 @@
-// Copyright: Ankitects Pty Ltd and contributors
-// License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-
-import { ChildNodeRange } from "./child-node-range";
-import { findAfter, findBefore } from "./find-adjacent";
-import { b, div, easyBold, easyItalic, i, t, u } from "./test-utils";
-
-describe("in a simple search", () => {
-    const before = b(t("before"));
-    const within = t("within");
-    const after = i(t("after"));
-    div(before, within, after);
-
-    let range: ChildNodeRange;
-
-    beforeEach(() => {
-        range = ChildNodeRange.fromNode(within);
-    });
-
-    describe("findBefore", () => {
-        test("finds an element", () => {
-            const matches = findBefore(range, easyBold.matcher);
-            expect([...range]).toEqual([before, within]);
-            expect(matches).toHaveLength(1);
-        });
-
-        test("does not find non-existing element", () => {
-            const matches = findBefore(range, easyItalic.matcher);
-            expect([...range]).toEqual([within]);
-            expect(matches).toHaveLength(0);
-        });
-    });
-
-    describe("findAfter", () => {
-        test("finds an element", () => {
-            const matches = findAfter(range, easyItalic.matcher);
-            expect([...range]).toEqual([within, after]);
-            expect(matches).toHaveLength(1);
-        });
-
-        test("does not find non-existing element", () => {
-            const matches = findAfter(range, easyBold.matcher);
-            expect([...range]).toEqual([within]);
-            expect(matches).toHaveLength(0);
-        });
-    });
-});
-
-describe("find by descension", () => {
-    describe("single top-level match", () => {
-        const before = u(b(t("before")));
-        const within = t("within");
-        const after = u(b(t("after")));
-        div(before, within, after);
-
-        let range: ChildNodeRange;
-
-        beforeEach(() => {
-            range = ChildNodeRange.fromNode(within);
-        });
-
-        test("findBefore", () => {
-            const matches = findBefore(range, easyBold.matcher);
-            expect([...range]).toEqual([before, within]);
-            expect(matches).toHaveLength(1);
-        });
-
-        test("findAfter", () => {
-            const matches = findAfter(range, easyBold.matcher);
-            expect([...range]).toEqual([within, after]);
-            expect(matches).toHaveLength(1);
-        });
-    });
-
-    describe("consecutive top-level", () => {
-        const before1 = u(b(t("before")));
-        const before2 = u(b(t("before")));
-        const within = t("within");
-        const after1 = u(b(t("after")));
-        const after2 = u(b(t("after")));
-        div(before1, before2, within, after1, after2);
-
-        let range: ChildNodeRange;
-
-        beforeEach(() => {
-            range = ChildNodeRange.fromNode(within);
-        });
-
-        test("findBefore", () => {
-            const matches = findBefore(range, easyBold.matcher);
-            expect([...range]).toEqual([before1, before2, within]);
-            expect(matches).toHaveLength(2);
-        });
-
-        test("findAfter", () => {
-            const matches = findAfter(range, easyBold.matcher);
-            expect([...range]).toEqual([within, after1, after2]);
-            expect(matches).toHaveLength(2);
-        });
-    });
-
-    describe("top-level consecutive within", () => {
-        const before = u(b(t("before")), b(t("before")));
-        const within = t("within");
-        const after = u(b(t("after")), b(t("after")));
-        div(before, within, after);
-
-        let range: ChildNodeRange;
-
-        beforeEach(() => {
-            range = ChildNodeRange.fromNode(within);
-        });
-
-        test("findBefore", () => {
-            const matches = findBefore(range, easyBold.matcher);
-            expect([...range]).toEqual([before, within]);
-            expect(matches).toHaveLength(2);
-        });
-
-        test("findAfter", () => {
-            const matches = findAfter(range, easyBold.matcher);
-            expect([...range]).toEqual([within, after]);
-            expect(matches).toHaveLength(2);
-        });
-    });
-
-    describe("no block-level", () => {
-        const within = t("within");
-        div(div(b(t("before"))), within, div(b(t("after"))));
-
-        let range: ChildNodeRange;
-
-        beforeEach(() => {
-            range = ChildNodeRange.fromNode(within);
-        });
-
-        test("findBefore", () => {
-            const matches = findBefore(range, easyBold.matcher);
-            expect([...range]).toEqual([within]);
-            expect(matches).toHaveLength(0);
-        });
-
-        test("findAfter", () => {
-            const matches = findAfter(range, easyBold.matcher);
-            expect([...range]).toEqual([within]);
-            expect(matches).toHaveLength(0);
-        });
-    });
-});
